Move to new mirage config style
The old route only style has been deprecated and we need to make a few
adjustments to keep moving forward.
