+++ package.json
@@ -105,7 +105,7 @@
     "ember-cli-dependency-lint": "^2.0.0",
     "ember-cli-deprecation-workflow": "^2.0.0",
     "ember-cli-inject-live-reload": "^2.1.0",
-    "ember-cli-mirage": "^2.3.1",
+    "ember-cli-mirage": "^2.4.0",
     "ember-cli-sass": "10.0.1",
     "ember-cli-server-variables": "^3.0.0",
     "ember-cli-sri": "^2.1.1",
