+++ tests/dummy/mirage/config.js
@@ -1,18 +1,26 @@
-import setupMirage from './setup';
+import commonRoutes from './routes';
 import ENV from 'dummy/config/environment';
+import { createServer, discoverEmberDataModels } from 'ember-cli-mirage';
 
 const { apiVersion } = ENV;
 
-export default function () {
-  this.namespace = '/';
-  setupMirage(this);
+export default function (config) {
+  let finalConfig = {
+    ...config,
+    models: { ...discoverEmberDataModels(), ...config.models },
+    routes() {
+      this.namespace = '/';
+      commonRoutes(this);
+      this.get('application/config', function () {
+        return {
+          config: {
+            type: 'form',
+            apiVersion,
+          },
+        };
+      });
+    },
+  };
 
-  this.get('application/config', function () {
-    return {
-      config: {
-        type: 'form',
-        apiVersion,
-      },
-    };
-  });
+  return createServer(finalConfig);
 }
